WIP on refactoring
